pyproject.toml
[project]
name = "image-editing-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai>=1.18.0",
    "numpy>=2.2.6",
    "opencv-python>=4.11.0.86",
    "pandas>=2.2.3",
    "piexif>=1.1.3",
    "pillow>=11.2.1",
    "pydantic>=2.11.5",
    "python-dotenv>=1.1.0",
    "sentence-transformers>=4.1.0",
    "tqdm>=4.67.1",
]
